[forge] compatibility test on cloud backend -fix client -introduce version and validator num -add swarm versions -add comp test -fixed image_tag incorrect -add full comp test -refactory cluster helper -add node recovery after update image -rebase on new forge cli -fix aws codebuild
Closes: #8888
Closes: #10614
