+++ docker/build-common.sh
@@ -50,7 +50,6 @@ if [ "$IMAGE_TARGETS" = "test" ] || [ "$IMAGE_TARGETS" = "all"  ]; then
           -p cluster-test \
           -p cli \
           -p diem-faucet \
-          -p forge \
           -p forge-cli \
           "$@"
 fi
