+++ testsuite/compatibility-test/src/lib.rs
@@ -5,7 +5,7 @@ use anyhow::bail;
 use diem_sdk::types::PeerId;
 use forge::{
     EmitJobRequest, NetworkContext, NetworkTest, NodeExt, Result, SwarmExt, Test, TxnEmitter,
-    Version,
+    TxnStats, Version,
 };
 use rand::SeedableRng;
 use std::time::{Duration, Instant};
@@ -20,7 +20,7 @@ fn batch_update<'t>(
         ctx.swarm().upgrade_validator(*validator, version)?;
     }
 
-    let deadline = Instant::now() + Duration::from_secs(5);
+    let deadline = Instant::now() + Duration::from_secs(60);
     for validator in validators_to_update {
         ctx.swarm()
             .validator_mut(*validator)
@@ -31,9 +31,12 @@ fn batch_update<'t>(
     Ok(())
 }
 
-fn generate_traffic<'t>(ctx: &mut NetworkContext<'t>, validators: &[PeerId]) -> Result<()> {
+pub fn generate_traffic<'t>(
+    ctx: &mut NetworkContext<'t>,
+    validators: &[PeerId],
+    duration: Duration,
+) -> Result<TxnStats> {
     let rt = Runtime::new()?;
-    let duration = Duration::from_secs(5);
     let rng = SeedableRng::from_rng(ctx.core().rng())?;
     let validator_clients = ctx
         .swarm()
@@ -42,10 +45,10 @@ fn generate_traffic<'t>(ctx: &mut NetworkContext<'t>, validators: &[PeerId]) -> 
         .map(|n| n.async_json_rpc_client())
         .collect::<Vec<_>>();
     let mut emitter = TxnEmitter::new(ctx.swarm().chain_info(), rng);
-    let _ =
+    let stats =
         rt.block_on(emitter.emit_txn_for(duration, EmitJobRequest::default(validator_clients)))?;
 
-    Ok(())
+    Ok(stats)
 }
 
 pub struct SimpleValidatorUpgrade;
@@ -83,7 +86,9 @@ impl NetworkTest for SimpleValidatorUpgrade {
         let mut first_batch = all_validators.clone();
         let second_batch = first_batch.split_off(first_batch.len() / 2);
         let first_node = first_batch.pop().unwrap();
+        let duration = Duration::from_secs(5);
 
+        println!("1. Downgrade all validators to older version");
         // Ensure that all validators are running the older version of the software
         let validators_to_downgrade = ctx
             .swarm()
@@ -94,26 +99,26 @@ impl NetworkTest for SimpleValidatorUpgrade {
         batch_update(ctx, &validators_to_downgrade, &old_version)?;
 
         // Generate some traffic
-        generate_traffic(ctx, &all_validators)?;
+        generate_traffic(ctx, &all_validators, duration)?;
 
         // Update the first Validator
-        println!("upgrading first Validator");
+        println!("2. upgrading first Validator");
         batch_update(ctx, &[first_node], &new_version)?;
-        generate_traffic(ctx, &[first_node])?;
+        generate_traffic(ctx, &[first_node], duration)?;
 
         // Update the rest of the first batch
-        println!("upgrading rest of first batch");
+        println!("3. upgrading rest of first batch");
         batch_update(ctx, &first_batch, &new_version)?;
-        generate_traffic(ctx, &first_batch)?;
+        generate_traffic(ctx, &first_batch, duration)?;
 
         ctx.swarm().fork_check()?;
 
         // Update the second batch
-        println!("upgrading second batch");
+        println!("4. upgrading second batch");
         batch_update(ctx, &second_batch, &new_version)?;
-        generate_traffic(ctx, &second_batch)?;
+        generate_traffic(ctx, &second_batch, duration)?;
 
-        println!("check swarm health");
+        println!("5. check swarm health");
         ctx.swarm().fork_check()?;
 
         Ok(())
